pyproject.toml
[tool.poetry]
name = "fs2es-indexer"
version = "0.4.1"
description = "Index files and directories into elastic search"
authors = ["Matthias Kühne <matthias.kuehne@ellerhold.de>"]
license = "proprietary"
readme = "README.md"
repository = "https://github.com/Ellerhold/fs2es-indexer"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = ">5.0"
elasticsearch = "^8.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
fs2es-indexer = "fs2es_indexer.fs2es_indexer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
